Handle the old pin widget somehow being closed already
This is a it of defensive coding that shouldn't normally be needed. I
was playing around with making the gui capture support specific regions
and managed to get the capture widget on a different screen to the pin.
Then closed the pin before finishing the new capture. Then the
existing.close() segfaulted because the reference to existing was
invalid.

It would have been easier to just do something like `existing.isValid()`
or `existing != nullptr` but apparently you can't check if a reference
is valid. Seems weird, but hence the extra signal juggling to deal with
this case.

New new bit is mostly closePinConn and destroyedConn (which just cleans
up the former). And I moved the "disconnect everything" to a seperate
variable to be able to re-use it.