When editing a pin, restrict gui window to just one screen
When editing a pinned image there is no need to cover all screens with the gui
window since the intention is that the user doesn't move the selection beyond
the pinned image. So it can be convenient for users with multiple monitors to
only have the monitor with the image being edited taken up by the gui window.

The main change here is in FlameshotDaemon. The positioning of the GUI windows
is defined by:

1) size of the screenshot pixmap fetched by the ScreenGrabber - this defines
   the size of the GUI window
2) a topLeft QPoint - this defines where on the desktop the GUI windows is
   located, so far it is always the top left point of the entire desktop

The daemon is changed in two ways if it sees a CaptureRequest with a capture
screen set on it:

1) it'll as the ScreenGrabber to get a pixmap just for that screen, instead of
for the whole desktop - this changes the GUI window size
2) it'll move the GUI window to the top left of only the requested screen

The only change to existing code that'll be executed without a capture screen
set in the request is the addition of `move(topLeft)` on Linux. mac and
windows are already doing their own moving. `topleft` will remain 0,0 in that
case so it should be no change in behaviour.

This probably doesn't currently work as intended on windows. That looks like
it is looping through all screens to redefine `topLeft`, I'm assuming 0,0 is
only relative to the primary screen or something on windows. I'll have to test
that and compare it to the corresponding code in ScreenGrabber.