refactor: remove fullscreen arg to CaptureWidget and minor cleanup
There is a `bool fullScreen` arg to CaptureWidget that wasn't being used
anywhere. There are a few places in the code which are conditional on that
variable but in fact the widget wouldn't be functional without those pieces of
code. This commit removes the argument to the initializer, the corresponding
attribute from the context attribute and any conditionals around them (the
code in the conditionals stayed).

I also shuffled stuff around a little bit in `initWindowFlags()` so that the
move and resize calls could be common between both paths that use them.