Fix help overlay positioning when drawing over a single screen
The overlay positions itself at the center of the passed area relative to the
parent widget. "Current Screen" here is the screen the cursor is under at the
time the screenshot flow starts.

So that works fine when the window is covering the whole desktop because the
absolute screen offset matches the offset relative to the parent. But when the
GUI window is covering just one screen we don't want to pass it the absolute
position of the screen otherwise if the screen is anywhere but at the top
left that absolute offset will be applied relative to the parent widget and
the overlay will be way off the screen.