Fix editing pin on single screen on Windows
This commit fixes the single screen gui window positioning on windows.
Previously it was always on the leftmost screen.

Previously windows would always calculate `topLeft` based on looking at the
topmost and leftmost points of any screens. Presumably because it can't rely
on 0,0 being the top left of the desktop? It would be interesting to compare
that to ScreenGrabber.desktopGeometry().

Anyway, when we only want to draw on a single window we either don't need to
go through all that effort or we should be combining the screen geometry with
the desktop geometry somehow. Like `ScreenGrabber.screenGeometry()` is doing.
Although that method is doing it in a wayland only block so I can't see how it
applies to windows...

This commit diff looks a bit messy but what it's doing is moving the windows
specific code into an else block inside #ifdefs.