wip: maybe fix windows initial wrong pin offset
On windows when you have two monitors and they aren't lined up at the
top perfectly, then when you pin a selection the pin window ends up
jumping a little bit. Specifically it seems to be half of the screen's
y-offset from the global 0?

Will this fix it? Not sure.

It seems some global measurement needs to be mapped to local? Maybe?