Examine m_context.selection too
On windows, when screen 1 is to the right of screen two, the offsets of
screen 2 are negative. When pinning a screenshot on screen one the width
of the pinned image somehow gets a whole screen width added to it. I'm
not sure where this comes from.

Debugging where the pin widget gets created from I see this stack:

    PinWidget::PinWidget
    FlameshotDaemon::attachPin
    FlameshotDaemon::createPin
    Flameshot::exportCapture
    CaptureWidget::~CaptureWidget

The extra width is evidenced in the initial geometry passed to the pin
widget. Although that doesn't explain why the pin widget is the correct
size...

Anyway, double checking that both selection variables in the capture
widget is worth doing.